# book-social-network-ui/src/app/services/models/history-book-response.ts
/* tslint:disable */
/* eslint-disable */
export interface HistoryBookResponse {
  authorName?: string;
  id?: number;
  isbn?: string;
  rate?: number;
  returnApproved?: boolean;
  returned?: boolean;
  title?: string;
}